requirements.txt
prometheus-client~=0.8.0
Flask~=1.1.2
Werkzeug~=1.0.1
waitress~=1.4.4
speedtest-cli~=2.1.2